pyproject.toml
[project]
name = "apex-ledger"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.30.0,<0.40",
    "asyncpg>=0.29.0,<0.30",
    "faker>=24.0.0,<26.0",
    "fastmcp>=0.9.0",
    "httpx==0.27.0",
    "langgraph>=0.1.0,<0.3",
    "mineru>=1.0",
    "openpyxl>=3.1.0,<4.0",
    "pydantic>=2.6.0,<3.0",
    "pytest>=8.0.0,<9.0",
    "pytest-asyncio>=0.23.0,<0.25",
    "python-dotenv>=1.0.0,<2.0",
    "reportlab>=4.2.0,<5.0",
]
